+++ lib/magiceye.dart
@@ -1,7 +1,8 @@
 library magiceye;
 
-export 'package:magiceye/contexts/control_layer_context.dart';
-export 'package:magiceye/contexts/preview_layer_context.dart';
-export 'package:magiceye/enums/device_direction.dart';
-export 'package:magiceye/layers/preview_layer.dart';
-export 'package:magiceye/widget/magiceye_widget.dart';
+export './src/contexts/control_layer_context.dart';
+export './src/contexts/preview_layer_context.dart';
+export './src/enums/device_direction.dart';
+export './src/widget/magiceye_widget.dart';
+export './src/layers/default_camera_control_layer.dart';
+export './src/layers/preview_layer.dart';